requirements.txt
Jinja2
python-docx
qrcode
Pillow
pytest
weasyprint